7k: Reduce the delay between two reads of command file
The native test application currently sleeps for two seconds if there is
nothing to read in the command file (/data/audio_test).

This delay is creating problem when test team tries to capture the
concurrency profile information in a repetitive fashion. The test
application doesn't provide the information continuously if task
information is captured while decoding very short files or while doing
non-tunneled mode file write operation of short duration files because
of this two seconds delay. Hence reducing this delay to 500 milliseconds

Change-Id: Id68b98f983ff71132e687db73aa861250bde738c