audio-native: 8660: Fix AC3 32Khz playback issue
-Problem: 32Khz AC3 clips are playing faster
-Cause: In hdmi_msm driver sample rate is set to 48Khz and this being
called during msm turn on. Sample rate cannot be set dynamically.
This resulting in fast playback when we are trying to play 32Khz clips.
-Fix:
1> Identify the sample rate before AUDIO_START. Call new API implemented
in parser library to get first frame information.
2> Change the ioctl AUDIO_SET_CONFIG to send msm_audio_config structure
with sample rate information and buffer size. AUDIO_SET_CONFIG ioctl
changes the sample rate of the hdmi driver based on sample rate send in
this structure

Change-Id: I5d297af45a890dfd913c007e4f0d933218f69660
CRs-Fixed: 317250, 317978